Approach to Employability Skills in Technical Education & its impact on satisfaction of selecting an Institute.
Ample of the current motivation to teach and assess employability skills will decide India’s ability to compete in the world economy. For a win-win situation, there must not be any differences in Institute’s perception and students’ experience in accessing employability skills for future prospective. What this paper aims to demonstrate is that, although as academics/service providers, we can use a general term “Employability skills”, we need to describe what is meant by this term to students, and colleges also need to recognize how their services and skill development are impacting on satisfaction of students while accessing employability skills. The purpose of this study is to examine the importance perceived by the institutes and experience realized by the students and to connect employability skills’ growth on students’ satisfaction of selection of an institute of Technical Educati