Promoting an Institute: An essential obligation for Technical Education Evolution
Purpose:Technical education in India contributes a major share to the overall education system and plays a
vital role in the social and economic development of the nation. Despite of rapid growth of technical
education in last decade in terms of the no. of institutes and intake capacity in India, institutes failed to
attract enrollments. Last three years saw, increasing gap in between the actual no. of enrollments and intake
capacity; 46% of seats were vacant in Technical Education in India in 2015. The purpose of this paper is to
highlight institute’s role in promoting enrollments in Technical education.
Design methodology: A qualitative research by a survey (through a structured questionnaire) of students
who are presently enrolled (Current-students) and those who have completed their study (Alumni) belonging
to the institutes offering Technical Education situated in Khandesh region of India and affiliated to the
N