ONLY EDUCATION IS NOT ENOUGH: A NECESSITY OF ALL-INCLUSIVE SERVICES FOR TECHNICAL EDUCATION.
Purpose: In addition to create challenging and congruent learning opportunities, it is important for Technical Education providers to create a supportive learning environment to help learners acclimate, connect with others and as well as access well-matched student support services to help them complete their studies and make them employable.Skill building is verycrucial to ensure employability of academia to understand and make sure good jobs. Unfortunately, most of the Technical Education institutes in India have not identified the importance of such services, resulting huge unemployment and dissatisfied students. The purpose of this paper is to highlight supporting and extension services in connection to operating services in Technical Education and students‘ perception on them.
Design methodology: A qualitative research by a survey (through a structured questionnaire) of students who are studying or have studi