Encouraging Enrolments by People’s Influence; A desperate need for Technical Education Evolution
There is a rapid growth of technical education in
last decade in terms of the no. of institutes and
intake capacity in India but with improper and nondirectional
development. As per AICTE reports,
there is noticeable gap in between the actual no. of
enrolments and intake capacity of these institutes. In
the year 2015-16, 46% of seats were vacant in
Technical Education in India. Segmentation and
‘People’ factor of service mix is very important
aspects in education services. The purpose of this
paper is to highlight the influence of people mix on
geographical and demographical factors of students
in selecting technical educational institute. The
study found that geographical and demographical
factors of students are highly related with the people
mix in selection of technical educational institute.
Several studies have revealed that the marketing
segmentation of technical education will be be