By The People For The People: People’s Influence On Students In Selecting Technical Educational Institute
Purpose: There is a rapid growth of technical education in last decade in terms of the no. of institutes and the intake capacity in India. However, there is noticeable gap in between the actual no. of enrollments and intake capacity of these institutes. In Maharashtra 41% of seats were vacant in 2016. ‘People’ factor of service mix is very important mix of education services. The purpose of this paper is to highlight the people factor that influence students in selecting technical educational institute of their choice.
Design methodology: A qualitative research was conducted with a survey of students those who are presently enrolled (pre-students) and those who have completed their study (post-students) belonging to the institutes offering technical education affiliated to the North Maharashtra University, Jalgaon.
Findings: The study found that there are some reference groups/influencers related