ARE THE SERVICES DELIVERED EMPLOYABLE? A SCENARIO OF TECHNICAL EDUCATION IN RURAL INDIA
Purpose
One of the most appalling challenges in India is persistently rising unemployment, explicitly in
the rural region. More than 20% of Indian youth between the ages of 15 and 24 years are “seeking
or available for work,” as per 2011 census data. There will be no peace and prosperity in the
country unless jobless people get appropriate channel. In India deficiency of skills is among the
main constraints for recruitment of Technical Education. Productivity of Technical Education in
not entirely absorbed due to lack of skills required by the employers. Along with the ample
arguments, taking place in recent times, regarding reformation in the productivity of Technical
Education, this present study syndicates focus on services delivered and its impact on employability
skills. The objective of this paper is to find out whether Technical Education institutes in rural part
are keen to focus on the employability