Downcast Enrollments: A Desperate Need of Holistic Marketing for Technical Education
Abstract: Purpose: Modern history has shown that only those countries, which could acquire capability to develop and apply science and technology, have found great success to grow their wealth and improve living conditions of their population. Technical education in India contributes a major share to the overall education system and plays a vital role in the social and economic development of the nation. There is a rapid growth of technical education in last decade in terms of the no. of institutes and intake capacity in India, however, institutes failed to attract enrollments which observed noticeable gap in between the actual no. of enrollments and intake capacity. In the year 2015-16, 46% of seats were vacant in Technical Education in India. The purpose of this paper is to highlight holistic marketing approach with promotion mix on diversified enrollments to motivate enrollments in selecting institute of Technical educati