e-Promotion: A Revolution In Technical Education Evolution
Despite of increasing pressure related to enrollment in Technical Education, some Institutes of Technical Education, successfully have taken on a corporate mentality to attract and retain high-quality students through their e-Promotion. Successful branding and marketing strategies have become most important for the Institutes of Technical Education to appeal to a growing and diversified student and to differentiate from competitors. Effective branding can be helpful in increasing enrollment and in developing relationships and bonding. This article explores internet based promotions that Technical Education institutes have focused on branding and marketing initiatives and its influence on the students in selecting institute. It also focuses on effectiveness of internet based promotions. The research study showed that Institute Website and Social Networking are the most influential promotion tools for making their decision of selection than the other t