Engineering a Woman: Marketing Opportunities and Challenges in India
Engineering plays a key role in supporting the growth and development of a country’s economy as well as in
improving the quality of life for citizens. Most of developed countries witnessed economic growth with the contribution of
women in engineering field. In developing countries like India women remained under-utilized resource. Women in
engineering are probably the single best investment that can be made in the developing world. There is clearly room for
improvement – not only in recruiting women into engineering, but also in retaining and promoting those women who wish and
do enter the profession. This study was designed to investigate a perception and determination of undergraduate women
towards accessing Engineering Education. The study has highlighted women’s perceptions and experiences on accessing
engineering education through institute’s Marketing Mix strategies which enables women to take up strategic positions to